refactor: subject storage file structure
Convert single file into directory module, with common and file-based code split into separate files. This was done because the previous storage file was getting too large for my taste
