+++ apps/ergo4all/lib/session_choice_screen.dart
@@ -9,7 +9,7 @@ import 'package:ergo4all/results/overview/screen.dart';
 import 'package:ergo4all/scenario/common.dart';
 import 'package:ergo4all/session_storage/session_storage.dart';
 import 'package:ergo4all/subjects/common.dart';
-import 'package:ergo4all/subjects/storage.dart';
+import 'package:ergo4all/subjects/storage/common.dart';
 import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
