+++ apps/ergo4all/lib/session_storage/src/fs.dart
@@ -7,7 +7,7 @@ import 'package:csv/csv.dart';
 import 'package:ergo4all/common/rula_session.dart';
 import 'package:ergo4all/scenario/common.dart';
 import 'package:ergo4all/session_storage/session_storage.dart';
-import 'package:ergo4all/subjects/storage.dart';
+import 'package:ergo4all/subjects/storage/common.dart';
 import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
