+++ apps/ergo4all/lib/subjects/subject_management_screen.dart
@@ -4,7 +4,7 @@ import 'package:common_ui/theme/styles.dart';
 import 'package:common_ui/widgets/red_circle_app_bar.dart';
 import 'package:ergo4all/subjects/common.dart';
 import 'package:ergo4all/subjects/creation/screen.dart';
-import 'package:ergo4all/subjects/storage.dart';
+import 'package:ergo4all/subjects/storage/common.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
